Revert "Merge pull request #27 from krongrah/SplashScreen"
This reverts commit b7bcc08deb87ed1b004fa5eba8188ccc95cefb5b, reversing
changes made to 43a02b4a75a9a5876ef647ab290091cebefe3853.
