+++ StarWars/Collision/src/main/java/game/collision/Collision.java
@@ -8,13 +8,13 @@ package game.collision;
 import common.Entity;
 import data.GameData;
 import data.World;
-import entityparts.LifePart;
 import entityparts.MovingPart;
 import entityparts.PositionPart;
 import entityparts.PropertiesPart;
 import enums.CollisionTypes;
+import static enums.CollisionTypes.SOLIDOBJECT;
 import enums.Directions;
-import interfaces.Projectile;
+import interfaces.Targetable;
 import org.openide.util.lookup.ServiceProvider;
 import services.IPostProcessor;
 
@@ -44,27 +44,24 @@ public class Collision implements IPostProcessor {
                 for (Entity target : world.getEntities()) {
                     PositionPart tarPos = target.getPart(PositionPart.class);
                     PropertiesPart tarProp = target.getPart(PropertiesPart.class);
-
-                    if (!object.equals(target)) {
-                        if (objPos.getY() >= (tarPos.getY() - (tarProp.getHeight() / 2)) && objPos.getY() <= (tarPos.getY() + (tarProp.getHeight() / 2))) {
-                            //System.out.println(((objProp.getWidth() / 2) - objPos.getX()) - ((tarProp.getWidth() / 2) + tarPos.getX()));
-                            if (objPos.getX() < tarPos.getX()) {
-                                //Check for right side collision exclusively
-                                //float dxR = (tarPos.getX() - objPos.getX()) - (tarProp.getWidth() / 2 - objProp.getWidth() / 2);
-                                //if (dxR < 0) {
-                                if ((tarPos.getX() - (tarProp.getWidth() / 2)) - ((objProp.getWidth() / 2) + objPos.getX()) <= 1) {
-                                    //Collision detected:
-                                    collide(object, target, Directions.RIGHT);
-                                }
-                            }
-                            if (objPos.getX() > tarPos.getX()) {
-                                //float dxL = (objPos.getX() - tarPos.getX()) - (objProp.getWidth() / 2 + tarProp.getWidth() / 2);
-                                //Check for left side collision exclusively
-                                if ((objPos.getX() - (objProp.getWidth() / 2)) - ((tarProp.getWidth() / 2) + tarPos.getX()) <= 1) {
-                                    collide(object, target, Directions.LEFT);
-                                }
+                    if (!object.equals(target) && tarProp.getCollisionType() == CollisionTypes.SOLIDOBJECT && objProp.getCollisionType() == CollisionTypes.SOLIDOBJECT) {
+                        if (objPos.getX() < tarPos.getX()) {
+                            //Check for right side collision exclusively                        
+                            float dxR = (tarPos.getX() - objPos.getX()) - (tarProp.getWidth() / 2 + objProp.getWidth() / 2);
+                            if (dxR < 0) {
+                                //Collision detected:
+                                collide(object, target, Directions.RIGHT);
                             }
                         }
+                        if (objPos.getX() > tarPos.getX()) {
+                            float dxL = (objPos.getX() - tarPos.getX()) - (objProp.getWidth() / 2 + tarProp.getWidth() / 2);
+                            //Check for left side collision exclusively
+                            if (dxL < 0) {
+                                //Collision detected:
+                                collide(object, target, Directions.LEFT);
+                            }
+                        }
+
                     }
                 }
             }
@@ -72,34 +69,100 @@ public class Collision implements IPostProcessor {
     }
 
     private void collide(Entity object, Entity target, Directions direction) {
-        //System.out.println("collide");
+        System.out.println("collide");
         PropertiesPart targetProp = target.getPart(PropertiesPart.class);
         PositionPart objectPos = object.getPart(PositionPart.class);
         PropertiesPart objectProp = object.getPart(PropertiesPart.class);
         PositionPart targetPos = target.getPart(PositionPart.class);
         MovingPart objectMov = object.getPart(MovingPart.class);
-        LifePart objLife = object.getPart(LifePart.class);
-
-        if (objectMov.isMoving()) {
-            if (objectProp.getCollisionType() == CollisionTypes.DAMAGE) {
-                //System.out.println("Object x: " + objectPos.getX() + " Target x: " + targetPos.getX());
-                if (object instanceof Projectile) {
-                    //It's a bullet
-                    objLife.setIsHit(true);
+        MovingPart targetMov = object.getPart(MovingPart.class);
+        if (targetProp.getCollisionType() == SOLIDOBJECT && objectMov.isMoving()) {
+//            if (targetMov.isMoving()) {
+//                //if both entities are moving
+//                
+//                //if the 
+//                float distance = (float) Math.abs(targetPos.getX() - objectPos.getX()) - targetProp.getWidth() / 2 - objectProp.getWidth() / 2;
+//                
+//                switch (direction) {
+//                    case RIGHT:
+//                        targetPos.setX(targetPos.getX() - distance / 2);
+//                        objectPos.setX(objectPos.getX() + distance / 2);
+//                        break;
+//                    case LEFT:
+//                        targetPos.setX(targetPos.getX() + distance / 2);
+//                        objectPos.setX(objectPos.getX() - distance / 2);
+//                        break;
+//                }
+//
+//            } else {
+                //if only the object is moving
+                System.out.println("single");
+                switch (direction) {
+                    case LEFT:
+                        objectPos.setX(targetPos.getX() + targetProp.getWidth() / 2 + objectProp.getWidth() / 2);
+                        break;
+                    case RIGHT:
+                        objectPos.setX(targetPos.getX() - targetProp.getWidth() / 2 - objectProp.getWidth() / 2);
+                        break;
                 }
-                if (target.hasPart(LifePart.class)) {
-                    LifePart tarLife = target.getPart(LifePart.class);
-                    tarLife.setIsHit(true);
-                }
-            }
-            switch (direction) {
-                case LEFT:
-                    objectPos.setX(targetPos.getX() + targetProp.getWidth() / 2 + objectProp.getWidth() / 2);
-                    break;
-                case RIGHT:
-                    objectPos.setX(targetPos.getX() - targetProp.getWidth() / 2 - objectProp.getWidth() / 2);
-                    break;
-            }
+            //}
         }
     }
+
 }
+
+//    /**
+//     * Sets a specific collision for the object only in regards to the y-axis
+//     *
+//     * @param object : Any Entity that moves
+//     * @param target : Any obstacle
+//     */
+//    private void setYAxisCollision(Entity object, Entity target) {
+//        PropertiesPart tarProp = target.getPart(PropertiesPart.class);
+//        PositionPart objPos = object.getPart(PositionPart.class);
+//        PositionPart tarPos = target.getPart(PositionPart.class);
+//
+//        object.setCollision(CollisionTypes.SOLIDOBJECT);
+//        if (tarProp.isObstacle()) {
+//            //Entity has bumped into obstacle on the y-axis:
+//            if (tarPos.getY() < objPos.getY()) {
+//                //Roof collision:
+//                object.setCollisionDirection(Directions.UP);
+//            } else {
+//                //Floor collision:
+//                object.setCollisionDirection(Directions.DOWN);
+//            }
+//
+//        }
+//    }
+//
+//    /**
+//     * Sets a specific collision for the object only in regards to the x-axis
+//     *
+//     * @param object : Any Entity that moves
+//     * @param target : Any obstacle
+//     */
+//    private void setXAxisCollision(Entity object, Entity target) {
+//        PropertiesPart tarProp = target.getPart(PropertiesPart.class);
+//        PositionPart objPos = object.getPart(PositionPart.class);
+//        PositionPart tarPos = target.getPart(PositionPart.class);
+//
+//        object.setCollision(CollisionTypes.SOLIDOBJECT);
+//        if (tarProp.isObstacle()) {
+//            //Entity has bumped into obstacle on the x-axis:
+//            if (tarPos.getX() < objPos.getX()) {
+//                //Right collision:
+//                object.setCollisionDirection(Directions.RIGHT);
+//            } else {
+//                //Left collision:
+//                object.setCollisionDirection(Directions.LEFT);
+//            }
+//        }
+//    }
+//
+//    private void resetEntityCollision(Entity object) {
+//        System.out.println("Resetting Collision");
+//        object.setCollision(CollisionTypes.NO_EFFECT);
+//        object.setCollisionDirection(null);
+//    }
+//}
