+++ StarWars/Common/src/main/java/common/Entity.java
@@ -6,6 +6,8 @@
 package common;
 
 import entityparts.EntityPart;
+import enums.CollisionTypes;
+import enums.Directions;
 import java.io.Serializable;
 import java.util.Map;
 import java.util.UUID;
@@ -18,7 +20,7 @@ import java.util.concurrent.ConcurrentHashMap;
 public class Entity implements Serializable {
 
     private final UUID ID = UUID.randomUUID();
-
+    
     private Map<Class, EntityPart> parts;
 
     public Entity() {
