+++ StarWars/Common/src/main/java/data/World.java
@@ -5,7 +5,6 @@
  */
 package data;
 
-import common.AISpawnPoint;
 import common.Entity;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -32,11 +31,6 @@ public class World {
     private final Map<String, Entity> entityMap = new ConcurrentHashMap<>();
 
     /**
-     * A list of all enemy spawn points read from the map file
-     */
-    private final List<AISpawnPoint> enemySpawnPoints = new ArrayList();
-
-    /**
      * Adds the given Entity to the world
      *
      * @param entity Is the entity to be added
@@ -45,14 +39,6 @@ public class World {
     public String addEntity(Entity entity) {
         entityMap.put(entity.getID(), entity);
         return entity.getID();
-    }
-
-    /**
-     * Adds a specific enemy spawn point to the list of spawn points
-     * @param spawn 
-     */
-    public void addEnemySpawn(AISpawnPoint spawn) {
-        enemySpawnPoints.add(spawn);
     }
 
     /**
@@ -71,10 +57,6 @@ public class World {
      */
     public Collection<Entity> getEntities() {
         return entityMap.values();
-    }
-    
-    public List<AISpawnPoint> getSpawnPoints(){
-        return this.enemySpawnPoints;
     }
 
     /**
