+++ StarWars/Core/src/main/java/game/core/Game.java
@@ -8,26 +8,85 @@ package game.core;
 import com.badlogic.gdx.ApplicationListener;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.GL20;
+import com.badlogic.gdx.graphics.OrthographicCamera;
+import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
+import common.Entity;
 import data.GameData;
 import data.World;
-import enums.State;
+import game.renderer.Renderer;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.concurrent.CopyOnWriteArrayList;
+import org.openide.util.Lookup;
+import org.openide.util.LookupListener;
+import org.openide.util.LookupEvent;
+import services.IProcessor;
+import services.IPluginService;
+import services.IPostProcessor;
+import services.IRenderer;
 
 /**
- * @author Andreas Bøgh Mølgaard-Andersen
+ *
  * @author ahmadhamid
  */
 public class Game implements ApplicationListener {
-    
-    private GameStateManager gameStateManager;
+
+    private static OrthographicCamera cam;
+    private ShapeRenderer sr;
+    private IRenderer renderer;
 
     private final GameData gameData = new GameData();
+    private final Lookup lookup = Lookup.getDefault();
+    private Lookup.Result<IPluginService> result;
+    private Lookup.Result<IPostProcessor> postProcessorResults;
+    private List<IPluginService> gamePlugins = new CopyOnWriteArrayList<>();
+    private List<IPostProcessor> postProcessors = new ArrayList<>();
+    private List<IProcessor> entityProcessors = new ArrayList<>();
 //    private List<IPluginService> entityPlugins = new ArrayList<>();
-    private final World world = new World();
+    private World world;
 
     @Override
     public void create() {
-        gameStateManager = new GameStateManager(gameData, world);
-        gameStateManager.setState(State.SPLASHSTATE); 
+
+        world = new World();
+        renderer = new Renderer(world);
+        renderer.setBackgroundColor(125, 190, 225, 1);
+
+        gameData.setDisplayWidth(Gdx.graphics.getWidth() * 2);
+        gameData.setDisplayHeight(Gdx.graphics.getHeight() * 2);
+
+        cam = new OrthographicCamera(gameData.getDisplayWidth(), gameData.getDisplayHeight());
+        cam.translate(gameData.getDisplayWidth(), gameData.getDisplayHeight());
+        cam.update();
+
+        sr = new ShapeRenderer();
+
+        result = lookup.lookupResult(IPluginService.class);
+        result.addLookupListener(lookupListener);
+        result.allItems();
+
+        Gdx.input.setInputProcessor(
+                new GameInputProcessor(gameData)
+        );
+
+//        IGamePluginService playerPlugin = new PlayerPlugin();
+//        IEntityProcessingService playerProcess = new PlayerControlSystem();
+//        entityPlugins.add(playerPlugin);
+//        entityProcessors.add(playerProcess);
+        // Lookup all Game Plugins using ServiceLoader
+        for (IPluginService iGamePlugin : result.allInstances()) {
+            iGamePlugin.start(gameData, world);
+            gamePlugins.add(iGamePlugin);
+            renderer.loadTexture(iGamePlugin.getPath());
+        }
+
+        postProcessorResults = lookup.lookupResult(IPostProcessor.class);
+        postProcessorResults.addLookupListener(lookupListener);
+        postProcessorResults.allItems();
+        for (IPostProcessor pp : postProcessorResults.allInstances()) {
+            postProcessors.add(pp);
+        }
     }
 
     @Override
@@ -42,26 +101,29 @@ public class Game implements ApplicationListener {
         update();
 
         draw();
-        
-        handleInput();
 
-//        gameData.getKeys().update();
-//        renderer.render(world, gameData);
+        gameData.getKeys().update();
+        renderer.render(world, gameData);
 
     }
 
     private void update() {
-        gameStateManager.update();
+        // Update
+        for (IProcessor entityProcessorService : getEntityProcessingServices()) {
+            entityProcessorService.process(gameData, world);
+        }
+        
+        for (IPostProcessor postProcessor : postProcessors) {
+            postProcessor.process(gameData, world);
+        }
     }
 
     private void draw() {
-        gameStateManager.draw();
+        for (Entity entity : world.getEntities()) {
+
+        }
     }
 
-    private void handleInput() {
-        gameStateManager.handleInput();
-    }
-    
     @Override
     public void resize(int width, int height) {
     }
@@ -76,7 +138,39 @@ public class Game implements ApplicationListener {
 
     @Override
     public void dispose() {
-        gameStateManager.dispose();
     }
+
+    private Collection<? extends IProcessor> getEntityProcessingServices() {
+        return lookup.lookupAll(IProcessor.class);
+//        return SPILocator.locateAll(IProcessor.class);
+    }
+
+//    private Collection<? extends IPostProcessor> getEntityPostProcessingServices() {
+//        return SPILocator.locateAll(IPostProcessor.class);
+//    }
+    private final LookupListener lookupListener = new LookupListener() {
+        @Override
+        public void resultChanged(LookupEvent le) {
+
+            Collection<? extends IPluginService> updated = result.allInstances();
+
+            for (IPluginService us : updated) {
+                // Newly installed modules
+                if (!gamePlugins.contains(us)) {
+                    us.start(gameData, world);
+                    gamePlugins.add(us);
+                }
+            }
+
+            // Stop and remove module
+            for (IPluginService gs : gamePlugins) {
+                if (!updated.contains(gs)) {
+                    gs.stop(gameData, world);
+                    gamePlugins.remove(gs);
+                }
+            }
+        }
+
+    };
 
 }
