+++ StarWars/Enemy/src/main/java/game/enemy/Enemy.java
@@ -5,7 +5,6 @@
  */
 package game.enemy;
 
-import enums.AITypes;
 import common.Entity;
 
 /**
@@ -13,28 +12,9 @@ import common.Entity;
  * @author andreasmolgaard-andersen
  */
 public class Enemy extends Entity {
-    private float originX;
-    private float originY;
-    private AITypes AIType;
+    
+    public Enemy() {
 
-    public Enemy(float originX, float originY) {
-        this.originX = originX;
-        this.originY = originY;
     }
-
-    public float getOriginX() {
-        return originX;
-    }
-
-    public float getOriginY() {
-        return originY;
-    }
-
-    public AITypes getAIType() {
-        return AIType;
-    }
-
-    public void setAIType(AITypes AIType) {
-        this.AIType = AIType;
-    }
+    
 }
