+++ StarWars/Enemy/src/main/java/game/enemy/EnemyPlugin.java
@@ -5,8 +5,6 @@
  */
 package game.enemy;
 
-import common.AISpawnPoint;
-import common.Entity;
 import data.GameData;
 import data.World;
 import entityparts.AnimationPart;
@@ -14,11 +12,10 @@ import entityparts.LifePart;
 import entityparts.MovingPart;
 import entityparts.PositionPart;
 import entityparts.PropertiesPart;
-import entityparts.WeaponPart;
-import static enums.AITypes.MELEE;
-import static enums.AITypes.SHOOTER;
+import enums.CollisionTypes;
 import org.openide.util.lookup.ServiceProvider;
 import services.IPluginService;
+import sprites.Sprites;
 
 @ServiceProvider(service = IPluginService.class)
 
@@ -27,50 +24,45 @@ import services.IPluginService;
  * @author andreasmolgaard-andersen
  */
 public class EnemyPlugin implements IPluginService {
-
-    private int shooterLife = 1;
-    private int meleeLife = 4;
-    private String enemyType;
-
+    
+    private Enemy enemy;
+    
     @Override
     public void start(GameData gameData, World world) {
-        for (AISpawnPoint spawnPoint : world.getSpawnPoints()) {
-            if (spawnPoint.getAIType() == SHOOTER) {
-                PositionPart enPos = spawnPoint.getPos();
-                Enemy enemy = new Enemy(enPos.getX(), enPos.getY());
-                enemy.add(enPos);
-                enemy.add(spawnPoint.getProp());
-                enemy.add(new LifePart(shooterLife));
-                enemy.add(new MovingPart(10, 100, 175));
-                enemy.add(new AnimationPart("TrooperIdle", 1, getPath()));
-                enemy.add(new WeaponPart());
-                enemy.setAIType(SHOOTER);
-                world.addEntity(enemy);
-            }
-            if (spawnPoint.getAIType() == MELEE) {
-                PositionPart enPos = spawnPoint.getPos();
-                Enemy enemy = new Enemy(enPos.getX(), enPos.getY());
-                enemy.add(enPos);
-                enemy.add(spawnPoint.getProp());
-                enemy.add(new LifePart(meleeLife));
-                enemy.add(new MovingPart(10, 100, 175));
-                enemy.add(new AnimationPart("RaiderIdle", 6, getPath()));
-                enemy.setAIType(MELEE);
-                world.addEntity(enemy);
-            }
-        }
+        enemy = new Enemy();
+        MovingPart mov=new MovingPart(10, 175, 250);
+        enemy.add(new PropertiesPart(32, 32, CollisionTypes.SOLIDOBJECT,false));
+        enemy.add(new LifePart(3));
+        enemy.add(new PositionPart(25*32, (39*32)+16));
+        enemy.add(mov);
+        enemy.add(new AnimationPart("Lukeidle", 5, getPath()));
+        //mov.setLeft(true);
+        
+        world.addEntity(enemy);
+        
+        
+        enemy = new Enemy();
+        
+        mov=new MovingPart(10, 100, 100);
+        
+        enemy.add(new PropertiesPart(32, 32, CollisionTypes.SOLIDOBJECT,false));
+        enemy.add(new LifePart(3));
+        enemy.add(new PositionPart(10*32, (39*32)+16));
+        enemy.add(mov);
+        enemy.add(new AnimationPart("Lukeidle", 5, getPath()));
+        //mov.setRight(true);
+        world.addEntity(enemy);
+        
     }
-
+    
     @Override
     public void stop(GameData gameData, World world) {
-        for (Entity entity : world.getEntities(Enemy.class)) {
-            world.removeEntity(entity);
-        }
+        world.removeEntity(enemy);
     }
 
     @Override
     public String getPath() {
-        return EnemyPlugin.class.getResource("/sprites/" + "Enemies" + ".txt").getPath();
+        return EnemyPlugin.class.getResource("/sprites/Luke.txt").getPath();
     }
-
+    
 }
