--- StarWars/Enemy/src/main/resources/sprites/Enemies.txt
@@ -1,194 +1,0 @@
-Enemies.png
-size: 192, 32
-format: RGBA8888
-filter: Linear,Linear
-repeat: none
-RaiderAttack0
-  rotate: false
-  xy: 0, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderAttack1
-  rotate: false
-  xy: 32, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderAttack2
-  rotate: false
-  xy: 64, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderAttack3
-  rotate: false
-  xy: 96, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderAttack4
-  rotate: false
-  xy: 128, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderAttack5
-  rotate: false
-  xy: 160, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderAttack6
-  rotate: false
-  xy: 192, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderIdle0
-  rotate: false
-  xy: 224, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderIdle1
-  rotate: false
-  xy: 256, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderIdle2
-  rotate: false
-  xy: 288, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderIdle3
-  rotate: false
-  xy: 320, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderIdle4
-  rotate: false
-  xy: 352, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderIdle5
-  rotate: false
-  xy: 384, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderIdle6
-  rotate: false
-  xy: 416, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderWalking0
-  rotate: false
-  xy: 448, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderWalking1
-  rotate: false
-  xy: 480, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderWalking2
-  rotate: false
-  xy: 512, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderWalking3
-  rotate: false
-  xy: 544, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderWalking4
-  rotate: false
-  xy: 576, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderWalking5
-  rotate: false
-  xy: 608, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-RaiderWalking6
-  rotate: false
-  xy: 640, 0
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-TrooperIdle0
-  rotate: false
-  xy: 0, 32
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-TrooperShooting0
-  rotate: false
-  xy: 32, 32
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-TrooperWalking0
-  rotate: false
-  xy: 64, 32
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-TrooperWalking1
-  rotate: false
-  xy: 96, 32
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-TrooperWalking2
-  rotate: false
-  xy: 128, 32
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
-TrooperWalking3
-  rotate: false
-  xy: 160, 32
-  size: 32, 32
-  orig: 32, 32
-  offset: 0, 0
-  index: -1
