+++ StarWars/Enemy/src/main/resources/sprites/Luke.txt
@@ -6,84 +6,84 @@ repeat: none
 Lukeidle0
   rotate: false
   xy: 0, 0
-  size: 22, 32
+  size: 32, 32
   orig: 32, 32
   offset: 0, 0
   index: -1
 Lukeidle1
   rotate: false
   xy: 32, 0
-  size: 22, 32
+  size: 32, 32
   orig: 32, 32
   offset: 0, 0
   index: -1
 Lukeidle2
   rotate: false
   xy: 64, 0
-  size: 22, 32
+  size: 32, 32
   orig: 32, 32
   offset: 0, 0
   index: -1
 Lukeidle3
   rotate: false
   xy: 96, 0
-  size: 22, 32
+  size: 32, 32
   orig: 32, 32
   offset: 0, 0
   index: -1
 Lukeidle4
   rotate: false
   xy: 128, 0
-  size: 22, 32
+  size: 32, 32
   orig: 32, 32
   offset: 0, 0
   index: -1
 Lukewalking0
   rotate: false
   xy: 160, 0
-  size: 22, 32
+  size: 32, 32
   orig: 32, 32
   offset: 0, 0
   index: -1
 Lukewalking1
   rotate: false
   xy: 192, 0
-  size: 22, 32
+  size: 32, 32
   orig: 32, 32
   offset: 0, 0
   index: -1
 Lukewalking2
   rotate: false
   xy: 224, 0
-  size: 22, 32
+  size: 32, 32
   orig: 32, 32
   offset: 0, 0
   index: -1
 Lukewalking3
   rotate: false
   xy: 256, 0
-  size: 22, 32
+  size: 32, 32
   orig: 32, 32
   offset: 0, 0
   index: -1
 Lukewalking4
   rotate: false
   xy: 288, 0
-  size: 22, 32
+  size: 32, 32
   orig: 32, 32
   offset: 0, 0
   index: -1
 Lukewalking5
   rotate: false
   xy: 320, 0
-  size: 22, 32
+  size: 32, 32
   orig: 32, 32
   offset: 0, 0
   index: -1
 Lukewalking6
   rotate: false
   xy: 352, 0
-  size: 22, 32
+  size: 32, 32
   orig: 32, 32
   offset: 0, 0
   index: -1
