+++ StarWars/Enemy/src/main/resources/sprites/test.txt
@@ -1,0 +1,19 @@
+test.png
+size: 64, 32
+format: RGBA8888
+filter: Linear,Linear
+repeat: none
+Blaster
+  rotate: false
+  xy: 0, 0
+  size: 32, 32
+  orig: 32, 32
+  offset: 0, 0
+  index: -1
+Saber
+  rotate: false
+  xy: 32, 0
+  size: 32, 32
+  orig: 32, 32
+  offset: 0, 0
+  index: -1
