+++ StarWars/Player/src/main/java/game/player/PlayerPlugin.java
@@ -17,6 +17,7 @@ import entityparts.WeaponPart;
 import enums.CollisionTypes;
 import org.openide.util.lookup.ServiceProvider;
 import services.IPluginService;
+import sprites.Sprites;
 
 @ServiceProvider(service = IPluginService.class)
 
@@ -34,10 +35,10 @@ public class PlayerPlugin implements IPluginService {
 
         player.add(new PropertiesPart(22, 32, CollisionTypes.SOLIDOBJECT,false));
         player.add(new LifePart(3));
-        player.add(new PositionPart(19*32, (39*32)+16));
+        player.add(new PositionPart(22*32, (39*32)+16));
         player.add(new MovingPart(10, 175, 250));
         player.add(new WeaponPart());
-        player.add(new AnimationPart("Lukeidle", 1, getPath()));
+        player.add(new AnimationPart("Lukeidle", 5, getPath()));
 
         world.addEntity(player);
     }
