+++ StarWars/Renderer/src/main/java/assetmanagement/AssetManagerClass.java
@@ -10,14 +10,19 @@ import com.badlogic.gdx.graphics.g2d.TextureAtlas;
  */
 public class AssetManagerClass {
 
+    //private String string;
     private AssetManager am;
+    //private String jarUrl = new File("").getAbsolutePath() + "/src/main/resources/sprites/Luke.png";
 
     public AssetManagerClass() {
+        //am = new AssetManager();
         am = new AssetManager(new AssetJarFileResolver());
     }
 
     public void Load(String path) {
         if (!path.equals("")) {
+            //string = path;
+
             am.load(path, TextureAtlas.class);
             while (am.update()) {
                 System.out.println("Asset loaded...");
