--- StarWars/Renderer/src/main/resources/sprites/SplashScreen.txt
@@ -1,82 +1,0 @@
-SplashScreen.png
-size: 7920, 407
-format: RGBA8888
-filter: Linear,Linear
-repeat: none
-SplashScreen0
-  rotate: false
-  xy: 0, 0
-  size: 720, 407
-  orig: 720, 407
-  offset: 0, 0
-  index: -1
-SplashScreen1
-  rotate: false
-  xy: 720, 0
-  size: 720, 407
-  orig: 720, 407
-  offset: 0, 0
-  index: -1
-SplashScreen2
-  rotate: false
-  xy: 1440, 0
-  size: 720, 407
-  orig: 720, 407
-  offset: 0, 0
-  index: -1
-SplashScreen3
-  rotate: false
-  xy: 2160, 0
-  size: 720, 407
-  orig: 720, 407
-  offset: 0, 0
-  index: -1
-SplashScreen4
-  rotate: false
-  xy: 2880, 0
-  size: 720, 407
-  orig: 720, 407
-  offset: 0, 0
-  index: -1
-SplashScreen5
-  rotate: false
-  xy: 3600, 0
-  size: 720, 407
-  orig: 720, 407
-  offset: 0, 0
-  index: -1
-SplashScreen6
-  rotate: false
-  xy: 4320, 0
-  size: 720, 407
-  orig: 720, 407
-  offset: 0, 0
-  index: -1
-SplashScreen7
-  rotate: false
-  xy: 5040, 0
-  size: 720, 407
-  orig: 720, 407
-  offset: 0, 0
-  index: -1
-SplashScreen8
-  rotate: false
-  xy: 5760, 0
-  size: 720, 407
-  orig: 720, 407
-  offset: 0, 0
-  index: -1
-SplashScreen9
-  rotate: false
-  xy: 6480, 0
-  size: 720, 407
-  orig: 720, 407
-  offset: 0, 0
-  index: -1
-SplashScreen10
-  rotate: false
-  xy: 7200, 0
-  size: 720, 407
-  orig: 720, 407
-  offset: 0, 0
-  index: -1
